Add option to download Mojang assets
Also fixes a bug with texture pack loading in the client.

Note: we cannot distribute Mojang assets, but some players might want to
use them. To that end, the launcher now has a button that allows the
user to download them him or herself. Since we download them directly
from Mojang, we're be good to go.